# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "azeezafeez212@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get filtered list of articles",
                "operationId": "93ded9e31119e6bbb22455c72ef97f19",
                "parameters": [
                    {
                        "name": "categories",
                        "in": "query",
                        "description": "Comma-separated list of categories (e.g., 'Business,Technology')",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sources",
                        "in": "query",
                        "description": "Comma-separated list of sources (e.g., 'guardian,newsapi')",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "authors",
                        "in": "query",
                        "description": "Comma-separated list of authors (e.g., 'John Doe,Jane Doe')",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Exact date filter (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Start date for range filtering (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "End date for range filtering (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query for title, content, category, source, or contributor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "categories": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Business"
                                                    }
                                                },
                                                "sources": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "guardian"
                                                    }
                                                },
                                                "authors": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{slug}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a single article by slug",
                "description": "Fetches a single article based on the provided slug.",
                "operationId": "getSingleArticle",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "source": {
                                                    "type": "string",
                                                    "example": "guardian"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "australia-news-live-pm-says-trump-tariffs-not"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Australia news live: PM says Trump tariffs ‘not a friendly act’"
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "example": "https://www.theguardian.com/australia-news/live/2025/mar/12/australia-news-live"
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "Australia news"
                                                },
                                                "contributor": {
                                                    "type": "string",
                                                    "example": "Stephanie Convery"
                                                },
                                                "published_at": {
                                                    "type": "string",
                                                    "example": "12th March, 2025 01:42"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "<div id='block-67d0e22b8f0879'>Content here</div>"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/filters": {
            "get": {
                "tags": [
                    "Filters"
                ],
                "summary": "Get list of article filter options",
                "operationId": "ac2fdf3b176313645fe5c885b613d133",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "categories": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "sources": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "authors": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "data": {
                                        "categories": [
                                            "Arts",
                                            "Business",
                                            "Technology",
                                            "Environment",
                                            "Health"
                                        ],
                                        "sources": [
                                            "guardian",
                                            "newsapi",
                                            "nytimes"
                                        ],
                                        "authors": [
                                            "Aashna Jain",
                                            "Amanda Taub",
                                            "Ben Casselman",
                                            "Ana Swanson",
                                            "Cade Metz"
                                        ]
                                    },
                                    "message": "Success"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Articles"
        },
        {
            "name": "Filters",
            "description": "Filters"
        }
    ]
}